+++ ui/src/mockState.tsx
@@ -21,6 +21,7 @@ export const localState: LocalStorage = {
       autoRefresh: 0,
       showTemplateControlBar: false,
       timeZone: 'Local' as TimeZone,
+      theme: 'dark',
     },
   },
   VERSION: '2.0.0',