+++ ui/src/shared/components/SingleStat.tsx
@@ -7,10 +7,12 @@ import {formatStatValue} from 'src/shared/utils/formatStatValue'
 
 // Types
 import {SingleStatViewProperties} from 'src/types/dashboards'
+import {Theme} from 'src/types'
 
 interface Props {
   properties: SingleStatViewProperties
   stat: number
+  theme: Theme
 }
 
 const SingleStat: SFC<Props> = ({stat, properties}) => {